Added Policies

+++ app/Http/Controllers/CategoryController.php
@@ -18,7 +18,7 @@ class CategoryController extends Controller
     }
 
     public function admin() {
-        $this->authorize('admin', User::class);
+        $this->authorize('create', User::class);
         return view('acp.knowledgebase',[
             'categories' => Category::all(),
             ]);
@@ -26,7 +26,7 @@ class CategoryController extends Controller
 
     public function categoryDetails(Request $request, int $id)
     {
-
+        $this->authorize('create', User::class);
         return view('acp.categorydetails', [
             'category' => Category::where( "catKey", $id )->first()
         ]);
@@ -34,10 +34,12 @@ class CategoryController extends Controller
     }
 
     public function newCategory () {
+        $this->authorize('create', User::class);
         return view('acp.newcategory');
     }
 
     public function acpEdit ( Request $request ) {
+        $this->authorize('update', Category::class);
         $request->validate([
             'category' => 'required',
             'status' => 'required',
@@ -55,7 +57,7 @@ class CategoryController extends Controller
     }
 
     public function acpDelete ( Request $request ) {
-
+        $this->authorize('delete', Category::class);
         DB::table('categories')->where('catKey', '=', $request->catKey )->delete();
 
         return view('acp.knowledgebase',[
@@ -64,6 +66,7 @@ class CategoryController extends Controller
     }
 
     public function acpSave ( Request $request ) {
+        $this->authorize('create', Category::class);
         Category::create([
             'catName' => $request->category,
             'catActive' => $request->status,
